refactor: 유저 인터페이스 제거
